live Video of model performance
Fabric masks are recommended to prevent onward transmission in the general population in public areas, particularly where distancing is not possible, and in areas of community transmission. This could include the school grounds in some situations. Masks may help to protect others, because wearers may be infected before symptoms of illness appear. The policy on wearing a mask or face covering should be in line with national or local guidelines. Where used, masks should be worn, cared for and disposed of properly.